requirements.txt
# Core packages
streamlit
pandas
openai
cohere
google-generativeai
re

# SQL connectors
mysql-connector-python
psycopg2-binary         # PostgreSQL
pyodbc                  # SQL Server (ODBC)
sqlite3                 # Built-in, no need to install

# LangChain
langchain
langchain-core
langchain-groq
langchain-google-genai

# Optional (for file path utilities)
pathlib



# streamlit==1.35.0
# pandas==2.2.2
# openai==1.32.0
# cohere==5.2.0
# google-generativeai==0.5.4
# mysql-connector-python==8.3.0
# psycopg2-binary==2.9.9
# pyodbc==5.1.0
# langchain==0.2.1
# langchain-core==0.2.2
# langchain-groq==0.0.3
# langchain-google-genai==0.0.5



# sqlite3
# pyodbc
# psycopg2
# langchain
# python-dotenv
# langchain-openai
# langchain-community
# mysql-connector-python
# openai
# langchain-groq
# cohere
# pandas
# streamlit